+++ src/api/consumer/lib/add.js
@@ -21,6 +21,7 @@ import { COMPONENT_ORIGINS, REGEX_PATTERN } from '../../../constants';
 import logger from '../../../logger/logger';
 import PathNotExists from './exceptions/path-not-exists';
 import EmptyDirectory from './exceptions/empty-directory';
+import type { ComponentMapFile } from '../../../consumer/bit-map/component-map';
 
 export default (async function addAction(
   componentPaths: string[],
@@ -38,7 +39,7 @@ export default (async function addAction(
   }
 
   // update test files according to dsl
-  async function getFiles(files, testFiles: string[]) {
+  async function getFiles(files: string[], testFiles: string[]): string[] {
     const fileList = testFiles.map(async (dsl) => {
       const fileList = files.map(async (file) => {
         const fileInfo = calculateFileInfo(file);
@@ -50,21 +51,25 @@ export default (async function addAction(
     });
     const fileListRes = R.flatten(await Promise.all(fileList));
     const uniq = R.uniq(fileListRes);
-    return uniq.map(file => pathNormalizeToLinux(file));
+    return uniq.map((testFile) => {
+      const relativeToConsumer = getPathRelativeToProjectRoot(testFile, consumer.getPath());
+      return pathNormalizeToLinux(relativeToConsumer);
+    });
   }
-  // todo: remove the logic of fixing the absolute paths, it is already done in BitMap class
+
   async function addOneComponent(componentPathsStats: Object, bitMap: BitMap, consumer: Consumer) {
     // remove excluded files from file list
     async function removeExcludedFiles(mapValues, excludedList) {
       const files = R.flatten(mapValues.map(x => x.files.map(i => i.relativePath)));
       const resolvedExcludedFiles = await getFiles(files, excludedList);
       mapValues.forEach((mapVal) => {
-        mapVal.files = mapVal.files.filter(key => !resolvedExcludedFiles.includes(key.relativePath));
+        if (resolvedExcludedFiles.includes(mapVal.mainFile)) { mapVal.files = []; } // if mainFile is excluded, exclude all files
+        else mapVal.files = mapVal.files.filter(key => !resolvedExcludedFiles.includes(key.relativePath));
       });
     }
 
     // used for updating main file if exists or doesn't exists
-    function addMainFileToFiles(files, mainFile) {
+    function addMainFileToFiles(files: ComponentMapFile[], mainFile) {
       if (mainFile && mainFile.match(REGEX_PATTERN)) {
         files.forEach((file) => {
           const fileInfo = calculateFileInfo(file.relativePath);
@@ -112,15 +117,15 @@ export default (async function addAction(
       updateIdAccordingToExistingComponent(id);
     }
 
-    async function getTestFiles(files, tests) {
+    async function getTestFiles(files: ComponentMapFile[]): ComponentMapFile[] {
       const testFilesArr = !R.isEmpty(tests) ? await getFiles(files.map(file => file.relativePath), tests) : [];
-      const testFiles = testFilesArr.map(testFile => ({
-        relativePath: pathNormalizeToLinux(testFile),
+      return testFilesArr.map(testFile => ({
+        relativePath: testFile,
         test: true,
         name: path.basename(testFile)
       }));
-      return testFiles;
     }
+
     const mapValuesP = await Object.keys(componentPathsStats).map(async (componentPath) => {
       if (componentPathsStats[componentPath].isDir) {
         const relativeComponentPath = getPathRelativeToProjectRoot(componentPath, consumer.getPath());
@@ -137,7 +142,7 @@ export default (async function addAction(
         });
 
         // get test files
-        const testFiles = await getTestFiles(files, tests);
+        const testFiles = await getTestFiles(files);
 
         const resolvedMainFile = addMainFileToFiles(files, pathNormalizeToLinux(main));
         // matches.forEach((match) => {
@@ -175,7 +180,7 @@ export default (async function addAction(
         { relativePath: pathNormalizeToLinux(relativeFilePath), test: false, name: path.basename(relativeFilePath) }
       ];
 
-      const testFiles = await getTestFiles(files, tests);
+      const testFiles = await getTestFiles(files);
       const resolvedMainFile = addMainFileToFiles(files, main);
       const mainFile = componentExists ? resolvedMainFile : relativeFilePath;
       return { componentId: parsedId, files: files.concat(testFiles), mainFile };