+++ src/cli/loader/loader-messages.js
@@ -10,6 +10,7 @@ export const BEFORE_RESET_ACTION = 'resetting component';
 export const BEFORE_IMPORT_ACTION = 'importing components';
 export const BEFORE_REMOTE_SEARCH = ({ scope, queryStr }: { scope: string, queryStr: string }) =>
   `searching remote scope <${scope}> for '${queryStr}'`;
+export const BEFORE_RUNNING_BUILD = 'building components';
 export const BEFORE_RUNNING_SPECS = 'running specs';
 export const BEFORE_IMPORT_PUT_ON_SCOPE = 'importing components';
 export const BEFORE_PERSISTING_PUT_ON_SCOPE = 'persisting...';