+++ src/scope/migrations/component-version/index.js
@@ -1,4 +1,11 @@
 import addBindingPrefixDeclartaion from './add-binding-prefix';
 import changeVersionToSemVerDeclartaion from './to-semver-migration';
+import implSpecsToFilesDeclartaion from './to-files';
+import specsResultstoArrayDeclartaion from './specs-results-to-array';
 
-export { addBindingPrefixDeclartaion, changeVersionToSemVerDeclartaion };
+export {
+  addBindingPrefixDeclartaion,
+  changeVersionToSemVerDeclartaion,
+  implSpecsToFilesDeclartaion,
+  specsResultstoArrayDeclartaion
+};