+++ src/scope/migrations/component-version/to-semver-migration.js
@@ -2,6 +2,8 @@
 import R from 'ramda';
 import semver from 'semver';
 import BitId from '../../../bit-id/bit-id';
+import logger from '../../../logger/logger';
+import { VERSION_DELIMITER } from '../../../constants';
 
 /**
  * Change the dependencies versions and the compiler / testers ids to semver
@@ -9,6 +11,14 @@ import BitId from '../../../bit-id/bit-id';
  */
 function changeVersionToSemVer(versionModel: Object): Object {
   const getUpdatedDependency = (dependency) => {
+    // Take care of very old models when the dependencies were strings
+    // in this case we will keep it string but change it to contain semver
+    // Those old model will still not work after this migration
+    if (typeof dependency === 'string') {
+      logger.warn("The dependency dependency is stored with an old format, this version won't work properly");
+      dependency = { id: _getUpdatedId(dependency), relativePaths: [] };
+      return dependency;
+    }
     dependency.id = _getUpdatedId(dependency.id);
     return dependency;
   };
@@ -28,14 +38,28 @@ function changeVersionToSemVer(versionModel: Object): Object {
 }
 
 function _getUpdatedId(id) {
-  const parsedId = BitId.parse(id);
-  let version = parsedId.getVersion().versionNum;
-  // In case there is already a semver, do nothing
-  if (!semver.valid(version)) {
-    version = `0.0.${version}`;
+  try {
+    // Fix also old version seperator (::)
+    if (id.includes('::')) {
+      id = id.replace('::', VERSION_DELIMITER);
+    }
+    const parsedId = BitId.parse(id);
+    // Don't convert latest word
+    if (parsedId.getVersion().latest) {
+      return id;
+    }
+
+    let version = parsedId.getVersion().versionNum;
+    // In case there is already a semver, do nothing
+    if (!semver.valid(version)) {
+      version = `0.0.${version}`;
+    }
+    const newId = `${parsedId.toStringWithoutVersion()}${VERSION_DELIMITER}${version}`;
+    return newId;
+  } catch (err) {
+    logger.error(`couldn't parse the id ${id} in order to migrate it to semver`);
+    throw err;
   }
-  const newId = `${parsedId.toStringWithoutVersion()}@${version}`;
-  return newId;
 }
 
 const changeVersionToSemVerDeclartaion = {