+++ src/scope/objects/repository.js
@@ -76,9 +76,14 @@ export default class Repository {
     const refs = await this.listRefs();
     return Promise.all(
       refs.map(async (ref) => {
-        const buffer = await this.loadRaw(ref);
-        const bitRawObject = await BitRawObject.fromDeflatedBuffer(buffer, ref.hash, this.types);
-        return bitRawObject;
+        try {
+          const buffer = await this.loadRaw(ref);
+          const bitRawObject = await BitRawObject.fromDeflatedBuffer(buffer, ref.hash, this.types);
+          return bitRawObject;
+        } catch (err) {
+          logger.error(`Couldn't load the ref ${ref} this object is probably corrupted and should be delete`);
+          return null;
+        }
       })
     );
   }
@@ -103,6 +108,12 @@ export default class Repository {
 
   loadRaw(ref: Ref): Promise<Buffer> {
     return readFile(this.objectPath(ref));
+  }
+
+  async loadRawObject(ref: Ref): Promise<BitRawObject> {
+    const buffer = await this.loadRaw(ref);
+    const bitRawObject = await BitRawObject.fromDeflatedBuffer(buffer, ref.hash, this.types);
+    return bitRawObject;
   }
 
   loadSync(ref: Ref, throws: boolean = true): BitObject {